requirements.txt
fasttext
transformers
torch
torchvision
torchaudio
numpy
pandas
moviepy
librosa
matplotlib
tqdm
scikit-learn
Pillow
opencv-python